==> 1.1_data_cpf.Rmd <==
---
title: "1.1_data_cpf"
author: "Jan Einhoff"
---

This R markdown file is used to prepare the baseline variables from the Comparative Panel File (CPF) for Germany (SOEP) and the UK (BHPS/UKHLS) for the analysis.

```{r setup}
# Load packages
library("tidyverse")
library("rio")
library("openxlsx")

# Clear workplace
rm(list = ls())

# Load raw comparative panel file (CPF) for UK and Germany
data_cpf <- import("data/data_raw/cpf/CPFv1.5.dta")

```

```{r}
# Select variables from CPF
data_cpf <- data_cpf %>% select(country, orgpid, hid, female, intyear, yborn, age, retf, oldpens, edu4, marstat5, kids_any, nphh, livpart, isco_1, public, selfemp, hhinc_post, disab, chron, srh5, sathlth5, migr, emplst6) %>% arrange(orgpid, age)

# Replace negative values
data_cpf[data_cpf < 0] <- NA

# Restrict to year range
data_cpf <- data_cpf %>% 
  filter(intyear %in% c(1991:2021))

# Restrict to age range and filter for respondents observed at age 50
data_cpf <- data_cpf %>% 
  filter(age %in% c(50:65)) %>% 
  group_by(orgpid) %>% 
  filter(min(age) == 50)

# Expand to full age range between age 50 and last age observed for each respondent
data_cpf <- data_cpf %>%
  group_by(orgpid) %>%
  mutate(min_age = min(age), max_age = max(age)) %>%
  complete(age = 50:65) %>%
  fill(c("orgpid", "hid", "min_age", "max_age", "country", "female", "yborn", "edu4", "migr"), .direction = "downup") %>%
  fill(c("kids_any", "isco_1", "public", "selfemp"), .direction = "down") %>%
  filter(age <= max_age & age >= min_age) %>%
  select(-c(min_age, max_age)) %>%
  ungroup()

```

```{r, warning = F}
# Modify base variables from CPF

# orgpid
data_cpf$orgpid <- as.character(data_cpf$orgpid)

# country
data_cpf <- data_cpf %>%
  mutate(country = ifelse(country == 6, "Germany", "UK")) %>%
  mutate(country = as.character(country))

# year
data_cpf <- data_cpf %>%
  mutate(year = as.numeric(intyear)) %>%
  select(-intyear)

# year_born
data_cpf <- data_cpf %>%
  mutate(year_born = as.numeric(yborn)) %>%
  select(-yborn)
  
# pension
data_cpf <- data_cpf %>%
  mutate(pension = as.numeric(oldpens)) %>%
  select(-oldpens)

# cohort_born
data_cpf <- data_cpf %>%
  mutate(cohort_born = case_when(
    year_born %in% c(1931:1935) ~ "1931:1935",
    year_born %in% c(1936:1940) ~ "1936:1940",
    year_born %in% c(1941:1945) ~ "1941:1945",
    year_born %in% c(1946:1950) ~ "1946:1950",
    year_born %in% c(1951:1955) ~ "1951:1955",
    year_born %in% c(1956:1960) ~ "1956:1960",
    year_born %in% c(1961:1965) ~ "1961:1965",
    year_born %in% c(1966:1970) ~ "1966:1970",
    year_born %in% c(1971:1975) ~ "1971:1975",
    year_born %in% c(1976:1980) ~ "1976:1980")) %>%
  mutate(cohort_born = as.character(cohort_born))

# migrant
data_cpf <- data_cpf %>%
  group_by(orgpid) %>%
  mutate(
    migrant = max(as.numeric(as.character(migr))),
    migrant = as.character(migrant)) %>%
  select(-migr)

# education_level
data_cpf <- data_cpf %>%
  group_by(orgpid) %>%
  mutate(
    edu4 = max(as.numeric(as.character(edu4)), na.rm = T),
    education_level = case_when(
    edu4 %in% c(1) ~ "Primary",
    edu4 %in% c(2, 3) ~ "Secondary",
    edu4 %in% c(4) ~ "Tertiary")) %>%
  mutate(education_level = as.character(education_level)) %>%
  select(-edu4)

# disability
data_cpf <- data_cpf %>%
  mutate(disability = as.character(disab)) %>%
  select(-disab)

# ill_health 
data_cpf <- data_cpf %>%
  mutate(poor_health_sb = case_when(
    sathlth5 %in% c(1, 2) ~ 1,
    sathlth5 %in% c(3, 4, 5) ~ 0),
    poor_health_sb = as.character(poor_health_sb)) %>%
  select(-sathlth5)

data_cpf <- data_cpf %>%
  mutate(poor_health_sr = case_when(
    srh5 %in% c(4, 5) ~ 1,
    srh5 %in% c(1, 2, 3) ~ 0),
    poor_health_sr = as.character(poor_health_sr)) %>%
  select(-srh5)

# marital_status
data_cpf <- data_cpf %>%
  mutate(marital_status = case_when(
    marstat5 %in% c(1) ~ "Married or Living with partner",
    marstat5 %in% c(2) ~ "Single",
    marstat5 %in% c(3) ~ "Widowed",
    marstat5 %in% c(4) ~ "Divorced or Separated")) %>%
  mutate(marital_status = as.character(marital_status)) %>%
  select(-marstat5)

# partner_in_hh
data_cpf <- data_cpf %>%
  mutate(partner_in_hh = as.character(livpart)) %>%
  select(-livpart)

# children_any
data_cpf <- data_cpf %>%
  mutate(children_any = as.character(kids_any)) %>%
  select(-kids_any)

# household_size
data_cpf <- data_cpf %>%
  mutate(household_size = as.numeric(nphh)) %>%
  select(-nphh)

# employment_status
data_cpf <- data_cpf %>%
  mutate(employment_status = case_when(
    emplst6 %in% c(1) ~ "Employed",
    emplst6 %in% c(2) ~ "Unemployed",
    emplst6 %in% c(3, 4, 5, 6) ~ "Inactive")) %>%
  mutate(employment_status = as.character(employment_status)) %>%
  select(-c(emplst6))

# retired
data_cpf <- data_cpf %>%
  group_by(orgpid) %>%
  mutate(retired = as.numeric(as.character(retf))) %>%
  fill(c("retired"), .direction = "down") %>%
  select(-retf)

# occupation_group
data_cpf <- data_cpf %>%
  mutate(occupation_group = case_when(
    isco_1 %in% c(1, 2) ~ "Managers and professionals",
    isco_1 %in% c(3) ~ "Technicians and associate professionals",
    isco_1 %in% c(4, 5) ~ "Clerks and service workers",
    isco_1 %in% c(0, 6, 9) ~ "Agricultural, elementary, and armed forces occupations",
    isco_1 %in% c(7, 8) ~ "Craft workers, machine operators and assemblers")) %>%
  mutate(occupation_group = as.character(occupation_group)) %>%
  select(-isco_1)

# public_sector
data_cpf <- data_cpf %>%
  mutate(public_sector = as.character(public)) %>%
  select(-public)

# self_employed
data_cpf <- data_cpf %>%
  mutate(self_employed = as.character(selfemp)) %>%
  select(-selfemp)

# hh_income_eq
data_cpf <- data_cpf %>%
  mutate(hh_income_eq = log((hhinc_post + 1) / sqrt(household_size))) %>%
  select(-hhinc_post)

# replace occupation, sector and self-employment for individuals outside the labour market
data_cpf <- data_cpf %>% mutate(
    occupation_group = ifelse(employment_status %in% c("Retired", "Inactive", "Unemployed"), "Out of LM", occupation_group),
    public_sector = ifelse(employment_status %in% c("Retired", "Inactive", "Unemployed"), "0", public_sector),
    self_employed = ifelse(employment_status %in% c("Retired", "Inactive", "Unemployed"), "0", self_employed))

```

```{r}
# add annual unemployment rate
unemp <- read.xlsx("data/data_raw/unemployment_rate.xlsx")

unemp <- unemp %>% mutate(year = as.numeric(year))

data_cpf <- data_cpf %>% mutate(year = year_born + age) %>% left_join(unemp, by = c("country", "year"))

# add annual inflation index
infl <- read.xlsx("data/data_raw/inflation_rates.xlsx")

infl <- infl %>% mutate(year = as.numeric(year))

data_cpf <- data_cpf %>% left_join(infl, by = c("country", "year"))

rm(unemp, infl)

```

```{r}
# gather variables and save
data_cpf <- data_cpf %>% select(country, orgpid, hid, age, year, female, year_born, cohort_born, retired, migrant, education_level, marital_status, partner_in_hh, children_any, household_size, poor_health_sr, poor_health_sb, employment_status, hh_income_eq, occupation_group, public_sector, self_employed, unemployment_rate, inflation_rate)

```

```{r}
save.image("data/data_preprocessed/data_cpf.RData")

```

```{r}
```
