==> 1.2_data_uk.Rmd <==
---
title: "1.2_data_de"
author: "Jan Einhoff"
---

This R markdown file adds missing variables from the UKHLS/BHPS to the base data set.

```{r setup}
# Load packages
library("tidyverse")
library("rio")

# Clear workplace
rm(list = ls())

# Load comparative panel file (CPF) for UK and Germany
load("data/data_preprocessed/data_cpf.RData")

# Filter for Germany and years used in the analysis 
data_uk <- data_cpf %>% filter(country == "UK")

rm(data_cpf)

```

# Social housing / Housing assistance
```{r}

```

# Partner employment status
```{r}
path_data <- "/Users/jlein/Documents/phd/current_activities/WP-5-House-ownership-retirement-timing/analyses/data/data_raw/ukhls/"

# Wave 1 (BHPS)
w1 <- import(paste0(path_data, "bhps/ba_indresp.dta"))
w1 <- w1 %>% select(pidp, ba_pno, ba_ppno, ba_hidp, ba_istrtdaty)
colnames(w1) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 2 (BHPS)
w2 <- import(paste0(path_data, "bhps/bb_indresp.dta"))
w2 <- w2 %>% select(pidp, bb_pno, bb_ppno, bb_hidp, bb_istrtdaty)
colnames(w2) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 3 (BHPS)
w3 <- import(paste0(path_data, "bhps/bc_indresp.dta"))
w3 <- w3 %>% select(pidp, bc_pno, bc_ppno, bc_hidp, bc_istrtdaty)
colnames(w3) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 4 (BHPS)
w4 <- import(paste0(path_data, "bhps/bd_indresp.dta"))
w4 <- w4 %>% select(pidp, bd_pno, bd_ppno, bd_hidp, bd_istrtdaty)
colnames(w4) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 5 (BHPS)
w5 <- import(paste0(path_data, "bhps/be_indresp.dta"))
w5 <- w5 %>% select(pidp, be_pno, be_ppno, be_hidp, be_istrtdaty)
colnames(w5) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 6 (BHPS)
w6 <- import(paste0(path_data, "bhps/bf_indresp.dta"))
w6 <- w6 %>% select(pidp, bf_pno, bf_ppno, bf_hidp, bf_istrtdaty)
colnames(w6) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 7 (BHPS)
w7 <- import(paste0(path_data, "bhps/bg_indresp.dta"))
w7 <- w7 %>% select(pidp, bg_pno, bg_ppno, bg_hidp, bg_istrtdaty)
colnames(w7) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 8 (BHPS)
w8 <- import(paste0(path_data, "bhps/bh_indresp.dta"))
w8 <- w8 %>% select(pidp, bh_pno, bh_ppno, bh_hidp, bh_istrtdaty)
colnames(w8) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 9 (BHPS)
w9 <- import(paste0(path_data, "bhps/bi_indresp.dta"))
w9 <- w9 %>% select(pidp, bi_pno, bi_ppno, bi_hidp, bi_istrtdaty)
colnames(w9) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 10 (BHPS)
w10 <- import(paste0(path_data, "bhps/bj_indresp.dta"))
w10 <- w10 %>% select(pidp, bj_pno, bj_ppno, bj_hidp, bj_istrtdaty)
colnames(w10) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 11 (BHPS)
w11 <- import(paste0(path_data, "bhps/bk_indresp.dta"))
w11 <- w11 %>% select(pidp, bk_pno, bk_ppno, bk_hidp, bk_istrtdaty)
colnames(w11) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 12 (BHPS)
w12 <- import(paste0(path_data, "bhps/bl_indresp.dta"))
w12 <- w12 %>% select(pidp, bl_pno, bl_ppno, bl_hidp, bl_istrtdaty)
colnames(w12) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 13 (BHPS)
w13 <- import(paste0(path_data, "bhps/bm_indresp.dta"))
w13 <- w13 %>% select(pidp, bm_pno, bm_ppno, bm_hidp, bm_istrtdaty)
colnames(w13) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 14 (BHPS)
w14 <- import(paste0(path_data, "bhps/bn_indresp.dta"))
w14 <- w14 %>% select(pidp, bn_pno, bn_ppno, bn_hidp, bn_istrtdaty)
colnames(w14) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 15 (BHPS)
w15 <- import(paste0(path_data, "bhps/bo_indresp.dta"))
w15 <- w15 %>% select(pidp, bo_pno, bo_ppno, bo_hidp, bo_istrtdaty)
colnames(w15) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 16 (BHPS)
w16 <- import(paste0(path_data, "bhps/bp_indresp.dta"))
w16 <- w16 %>% select(pidp, bp_pno, bp_ppno, bp_hidp, bp_istrtdaty)
colnames(w16) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 17 (BHPS)
w17 <- import(paste0(path_data, "bhps/bq_indresp.dta"))
w17 <- w17 %>% select(pidp, bq_pno, bq_ppno, bq_hidp, bq_istrtdaty)
colnames(w17) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 18 (BHPS)
w18 <- import(paste0(path_data, "bhps/br_indresp.dta"))
w18 <- w18 %>% select(pidp, br_pno, br_ppno, br_hidp, br_istrtdaty)
colnames(w18) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 19 (UKHLS)
w19 <- import(paste0(path_data, "ukhls/a_indresp.dta"))
w19 <- w19 %>% select(pidp, a_pno, a_ppno, a_hidp, a_intdaty_dv)
colnames(w19) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 20 (UKHLS)
w20 <- import(paste0(path_data, "ukhls/b_indresp.dta"))
w20 <- w20 %>% select(pidp, b_pno, b_ppno, b_hidp, b_intdaty_dv)
colnames(w20) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 21 (UKHLS)
w21 <- import(paste0(path_data, "ukhls/c_indresp.dta"))
w21 <- w21 %>% select(pidp, c_pno, c_ppno, c_hidp, c_intdaty_dv)
colnames(w21) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 22 (UKHLS)
w22 <- import(paste0(path_data, "ukhls/d_indresp.dta"))
w22 <- w22 %>% select(pidp, d_pno, d_ppno, d_hidp, d_intdaty_dv)
colnames(w22) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 23 (UKHLS)
w23 <- import(paste0(path_data, "ukhls/e_indresp.dta"))
w23 <- w23 %>% select(pidp, e_pno, e_ppno, e_hidp, e_intdaty_dv)
colnames(w23) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 24 (UKHLS)
w24 <- import(paste0(path_data, "ukhls/f_indresp.dta"))
w24 <- w24 %>% select(pidp, f_pno, f_ppno, f_hidp, f_intdaty_dv)
colnames(w24) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 25 (UKHLS)
w25 <- import(paste0(path_data, "ukhls/g_indresp.dta"))
w25 <- w25 %>% select(pidp, g_pno, g_ppno, g_hidp, g_intdaty_dv)
colnames(w25) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 26 (UKHLS)
w26 <- import(paste0(path_data, "ukhls/h_indresp.dta"))
w26 <- w26 %>% select(pidp, h_pno, h_ppno, h_hidp, h_intdaty_dv)
colnames(w26) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 27 (UKHLS)
w27 <- import(paste0(path_data, "ukhls/i_indresp.dta"))
w27 <- w27 %>% select(pidp, i_pno, i_ppno, i_hidp, i_intdaty_dv)
colnames(w27) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 28 (UKHLS)
w28 <- import(paste0(path_data, "ukhls/j_indresp.dta"))
w28 <- w28 %>% select(pidp, j_pno, j_ppno, j_hidp, j_intdaty_dv)
colnames(w28) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 29 (UKHLS)
w29 <- import(paste0(path_data, "ukhls/k_indresp.dta"))
w29 <- w29 %>% select(pidp, k_pno, k_ppno, k_hidp, k_intdaty_dv)
colnames(w29) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Wave 30 (UKHLS)
w30 <- import(paste0(path_data, "ukhls/l_indresp.dta"), encoding = "UTF-8")
w30 <- w30 %>% select(pidp, l_pno, l_ppno, l_hidp, l_intdaty_dv)
colnames(w30) <- c("orgpid", "person_n", "partner_n", "hid", "year")

# Combine
waves <- rbind(w1, w2, w3, w4, w5, w6, w7, w8, w9, w10, w12, w13, w14, w15, w16, w17, w18, w19, w20, w21, w22, w23, w24, w25, w26, w27, w28, w29, w30)

rm(path_data, w1, w2, w3, w4, w5, w6, w7, w8, w9, w10, w11, w12, w13, w14, w15, w16, w17, w18, w19, w20, w21, w22, w23, w24, w25, w26, w27, w28, w29, w30)

waves <- waves %>% unique()

waves$year <- as.numeric(waves$year)
waves$orgpid <- as.character(waves$orgpid)

# Add partner id to main dataset
data_uk <- data_uk %>% select(-hid) %>% unique() %>% left_join(waves %>% select(orgpid, hid, year, partner_n) %>% unique(), by = c("orgpid", "year"))

# Merge with main
partner_employment <- import("data/data_raw/cpf/CPFv1.5.dta")

partner_employment <- partner_employment %>% 
  filter(country == 7) %>% 
  select(country, age, orgpid, female, intyear, emplst6) %>% 
  rename("year" = "intyear") %>% 
  mutate(orgpid = as.character(orgpid), year = as.numeric(year)) %>%
  unique() %>% group_by(orgpid, year) %>% slice(1)

partner_employment <- partner_employment %>% unique() %>% left_join(waves, by = c("orgpid", "year"))

partner_employment <- partner_employment %>% ungroup() %>% select(hid, year, person_n, emplst6) %>% rename("employment_status_partner" = "emplst6") %>% na.omit()

partner_employment <- partner_employment %>%
  mutate(employment_status_partner = case_when(
    employment_status_partner %in% c(1) ~ "Employed",
    employment_status_partner %in% c(2) ~ "Unemployed",
    employment_status_partner %in% c(3, 4, 5, 6) ~ "Inactive")) %>%
  mutate(employment_status_partner = as.character(employment_status_partner)) %>%
  unique() %>% group_by(hid, year, person_n) %>% slice(1)

# Add partner employment status to main dataset
data_uk <- data_uk %>% unique() %>% left_join(partner_employment, by = c("hid", "year", "partner_n" = "person_n"))

data_uk <- data_uk %>% 
  mutate(employment_status_partner = ifelse(partner_n == 0, "No partner", employment_status_partner)) %>%
  mutate(employment_status_partner = ifelse(marital_status != "Married or Living with partner", "No partner", employment_status_partner)) %>%
  select(-c(hid, partner_n))

rm(partner_employment, waves)

```

# Housing and job tenure, region
```{r}
path_data <- "data/data_raw/ukhls/"

# Wave 1 (BHPS)
w1 <- import(paste0(path_data, "bhps/ba_indresp.dta"))

w1 <- w1 %>% select(pidp, ba_hidp, ba_age_dv, ba_tenure_dv, ba_jbbgy, ba_gor_dv, ba_jlendy) %>% mutate(across(everything(), as.character))

colnames(w1) <- c("orgpid", "hid", "age", "homeowner", "job_tenure", "region", "last_job")

# Wave 2 (BHPS)
w2 <- import(paste0(path_data, "bhps/bb_indresp.dta"))

w2 <- w2 %>% select(pidp, bb_hidp, bb_age_dv, bb_tenure_dv, bb_jbbgy, bb_gor_dv, bb_jlendy) %>% mutate(across(everything(), as.character))

colnames(w2) <- c("orgpid", "hid", "age", "homeowner", "job_tenure", "region", "last_job")

# Wave 3 (BHPS)
w3 <- import(paste0(path_data, "bhps/bc_indresp.dta"))

w3 <- w3 %>% select(pidp, bc_hidp, bc_age_dv, bc_tenure_dv, bc_jbbgy, bc_gor_dv, bc_jlendy) %>% mutate(across(everything(), as.character))

colnames(w3) <- c("orgpid", "hid", "age", "homeowner", "job_tenure", "region", "last_job")

# Wave 4 (BHPS)
w4 <- import(paste0(path_data, "bhps/bd_indresp.dta"))

w4 <- w4 %>% select(pidp, bd_hidp, bd_age_dv, bd_tenure_dv, bd_jbbgy, bd_gor_dv, bd_jlendy) %>% mutate(across(everything(), as.character))

colnames(w4) <- c("orgpid", "hid", "age", "homeowner", "job_tenure", "region", "last_job")

# Wave 5 (BHPS)
w5 <- import(paste0(path_data, "bhps/be_indresp.dta"))

w5 <- w5 %>% select(pidp, be_hidp, be_age_dv, be_tenure_dv, be_jbbgy, be_gor_dv, be_jlendy) %>% mutate(across(everything(), as.character))

colnames(w5) <- c("orgpid", "hid", "age", "homeowner", "job_tenure", "region", "last_job")

# Wave 6 (BHPS)
w6 <- import(paste0(path_data, "bhps/bf_indresp.dta"))

w6 <- w6 %>% select(pidp, bf_hidp, bf_age_dv, bf_tenure_dv, bf_jbbgy, bf_gor_dv, bf_jlendy) %>% mutate(across(everything(), as.character))

colnames(w6) <- c("orgpid", "hid", "age", "homeowner", "job_tenure", "region", "last_job")

# Wave 7 (BHPS)
w7 <- import(paste0(path_data, "bhps/bg_indresp.dta"))

w7 <- w7 %>% select(pidp, bg_hidp, bg_age_dv, bg_tenure_dv, bg_jbbgy, bg_gor_dv, bg_jlendy) %>% mutate(across(everything(), as.character))

colnames(w7) <- c("orgpid", "hid", "age", "homeowner", "job_tenure", "region", "last_job")

# Wave 8 (BHPS)
w8 <- import(paste0(path_data, "bhps/bh_indresp.dta"))

w8 <- w8 %>% select(pidp, bh_hidp, bh_age_dv, bh_tenure_dv, bh_jbbgy, bh_gor_dv, bh_jlendy) %>% mutate(across(everything(), as.character))

colnames(w8) <- c("orgpid", "hid", "age", "homeowner", "job_tenure", "region", "last_job")

# Wave 9 (BHPS)
w9 <- import(paste0(path_data, "bhps/bi_indresp.dta"))

w9 <- w9 %>% select(pidp, bi_hidp, bi_age_dv, bi_tenure_dv, bi_jbbgy, bi_gor_dv, bi_jlendy) %>% mutate(across(everything(), as.character))

colnames(w9) <- c("orgpid", "hid", "age", "homeowner", "job_tenure", "region", "last_job")

# Wave 10 (BHPS)
w10 <- import(paste0(path_data, "bhps/bj_indresp.dta"))

w10 <- w10 %>% select(pidp, bj_hidp, bj_age_dv, bj_tenure_dv, bj_jbbgy, bj_gor_dv, bj_jlendy) %>% mutate(across(everything(), as.character))

colnames(w10) <- c("orgpid", "hid", "age", "homeowner", "job_tenure", "region", "last_job")

# Wave 11 (BHPS)
w11 <- import(paste0(path_data, "bhps/bk_indresp.dta"))

w11 <- w11 %>% select(pidp, bk_hidp, bk_age_dv, bk_tenure_dv, bk_jbbgy, bk_gor_dv, bk_jlendy) %>% mutate(across(everything(), as.character))

colnames(w11) <- c("orgpid", "hid", "age", "homeowner", "job_tenure", "region", "last_job")

# Wave 12 (BHPS)
w12 <- import(paste0(path_data, "bhps/bl_indresp.dta"))

w12 <- w12 %>% select(pidp, bl_hidp, bl_age_dv, bl_tenure_dv, bl_jbbgy, bl_gor_dv, bl_jlendy) %>% mutate(across(everything(), as.character))

colnames(w12) <- c("orgpid", "hid", "age", "homeowner", "job_tenure", "region", "last_job")

# Wave 13 (BHPS)
w13 <- import(paste0(path_data, "bhps/bm_indresp.dta"))

w13 <- w13 %>% select(pidp, bm_hidp, bm_age_dv, bm_tenure_dv, bm_jbbgy, bm_gor_dv, bm_jlendy) %>% mutate(across(everything(), as.character))

colnames(w13) <- c("orgpid", "hid", "age", "homeowner", "job_tenure", "region", "last_job")

# Wave 14 (BHPS)
w14 <- import(paste0(path_data, "bhps/bn_indresp.dta"))

w14 <- w14 %>% select(pidp, bn_hidp, bn_age_dv, bn_tenure_dv, bn_jbbgy, bn_gor_dv, bn_jlendy) %>% mutate(across(everything(), as.character))

colnames(w14) <- c("orgpid", "hid", "age", "homeowner", "job_tenure", "region", "last_job")

# Wave 15 (BHPS)
w15 <- import(paste0(path_data, "bhps/bo_indresp.dta"))

w15 <- w15 %>% select(pidp, bo_hidp, bo_age_dv, bo_tenure_dv, bo_jbbgy, bo_gor_dv, bo_jlendy) %>% mutate(across(everything(), as.character))

colnames(w15) <- c("orgpid", "hid", "age", "homeowner", "job_tenure", "region", "last_job")

# Wave 16 (BHPS)
w16 <- import(paste0(path_data, "bhps/bp_indresp.dta"))

w16 <- w16 %>% select(pidp, bp_hidp, bp_age_dv, bp_tenure_dv, bp_jbbgy, bp_gor_dv, bp_jlendy) %>% mutate(across(everything(), as.character))

colnames(w16) <- c("orgpid", "hid", "age", "homeowner", "job_tenure", "region", "last_job")

# Wave 17 (BHPS)
w17 <- import(paste0(path_data, "bhps/bq_indresp.dta"))

w17 <- w17 %>% select(pidp, bq_hidp, bq_age_dv, bq_tenure_dv, bq_jbbgy, bq_gor_dv, bq_jlendy) %>% mutate(across(everything(), as.character))

colnames(w17) <- c("orgpid", "hid", "age", "homeowner", "job_tenure", "region", "last_job")

# Wave 18 (BHPS)
w18 <- import(paste0(path_data, "bhps/br_indresp.dta"))

w18 <- w18 %>% select(pidp, br_hidp, br_age_dv, br_tenure_dv, br_jbbgy, br_gor_dv, br_jlendy) %>% mutate(across(everything(), as.character))

colnames(w18) <- c("orgpid", "hid", "age", "homeowner", "job_tenure", "region", "last_job")

# Wave 19 (UKHLS)
w19 <- import(paste0(path_data, "ukhls/a_hhresp.dta"))
w19 <- w19 %>% select(a_hidp, a_tenure_dv) %>% mutate(across(everything(), as.character))

w19_ind <- import(paste0(path_data, "ukhls/a_indresp.dta"))
w19_ind <- w19_ind %>% select(pidp, a_hidp, a_age_dv, a_jbbgy, a_gor_dv, a_jlendy) %>% mutate(across(everything(), as.character))

w19 <- w19 %>% left_join(w19_ind, by = "a_hidp")

colnames(w19) <- c("hid", "homeowner", "orgpid", "age", "job_tenure", "region", "last_job")

rm(w19_ind)

# Wave 20 (UKHLS)
w20 <- import(paste0(path_data, "ukhls/b_hhresp.dta"))
w20 <- w20 %>% select(b_hidp, b_tenure_dv) %>% mutate(across(everything(), as.character))

w20_ind <- import(paste0(path_data, "ukhls/b_indresp.dta"))
w20_ind <- w20_ind %>% select(pidp, b_hidp, b_age_dv, b_jbbgy, b_gor_dv, b_jlendy) %>% mutate(across(everything(), as.character))

w20 <- w20 %>% left_join(w20_ind, by = "b_hidp")

colnames(w20) <- c("hid", "homeowner", "orgpid", "age", "job_tenure", "region", "last_job")

rm(w20_ind)

# Wave 21 (UKHLS)
w21 <- import(paste0(path_data, "ukhls/c_hhresp.dta"))
w21 <- w21 %>% select(c_hidp, c_tenure_dv) %>% mutate(across(everything(), as.character))

w21_ind <- import(paste0(path_data, "ukhls/c_indresp.dta"))
w21_ind <- w21_ind %>% select(pidp, c_hidp, c_age_dv, c_jbbgy, c_gor_dv, c_jlendy) %>% mutate(across(everything(), as.character))

w21 <- w21 %>% left_join(w21_ind, by = "c_hidp")

colnames(w21) <- c("hid", "homeowner", "orgpid", "age", "job_tenure", "region", "last_job")

rm(w21_ind)

# Wave 22 (UKHLS)
w22 <- import(paste0(path_data, "ukhls/d_hhresp.dta"))
w22 <- w22 %>% select(d_hidp, d_tenure_dv) %>% mutate(across(everything(), as.character))

w22_ind <- import(paste0(path_data, "ukhls/d_indresp.dta"))
w22_ind <- w22_ind %>% select(pidp, d_hidp, d_age_dv, d_jbbgy, d_gor_dv, d_jlendy) %>% mutate(across(everything(), as.character))

w22 <- w22 %>% left_join(w22_ind, by = "d_hidp")

colnames(w22) <- c("hid", "homeowner", "orgpid", "age", "job_tenure", "region", "last_job")

rm(w22_ind)

# Wave 23 (UKHLS)
w23 <- import(paste0(path_data, "ukhls/e_hhresp.dta"))
w23 <- w23 %>% select(e_hidp, e_tenure_dv) %>% mutate(across(everything(), as.character))

w23_ind <- import(paste0(path_data, "ukhls/e_indresp.dta"))
w23_ind <- w23_ind %>% select(pidp, e_hidp, e_age_dv, e_jbbgy, e_gor_dv, e_jlendy) %>% mutate(across(everything(), as.character))

w23 <- w23 %>% left_join(w23_ind, by = "e_hidp")

colnames(w23) <- c("hid", "homeowner", "orgpid", "age", "job_tenure", "region", "last_job")

rm(w23_ind)

# Wave 24 (UKHLS)
w24 <- import(paste0(path_data, "ukhls/f_hhresp.dta"))
w24 <- w24 %>% select(f_hidp, f_tenure_dv) %>% mutate(across(everything(), as.character))

w24_ind <- import(paste0(path_data, "ukhls/f_indresp.dta"))
w24_ind <- w24_ind %>% select(pidp, f_hidp, f_age_dv, f_jbbgy, f_gor_dv, f_jlendy) %>% mutate(across(everything(), as.character))

w24 <- w24 %>% left_join(w24_ind, by = "f_hidp")

colnames(w24) <- c("hid", "homeowner", "orgpid", "age", "job_tenure", "region", "last_job")

rm(w24_ind)

# Wave 25 (UKHLS)
w25 <- import(paste0(path_data, "ukhls/g_hhresp.dta"))
w25 <- w25 %>% select(g_hidp, g_tenure_dv) %>% mutate(across(everything(), as.character))

w25_ind <- import(paste0(path_data, "ukhls/g_indresp.dta"))
w25_ind <- w25_ind %>% select(pidp, g_hidp, g_age_dv, g_jbbgy, g_gor_dv, g_jlendy) %>% mutate(across(everything(), as.character))

w25 <- w25 %>% left_join(w25_ind, by = "g_hidp")

colnames(w25) <- c("hid", "homeowner", "orgpid", "age", "job_tenure", "region", "last_job")

rm(w25_ind)

# Wave 26 (UKHLS)
w26 <- import(paste0(path_data, "ukhls/h_hhresp.dta"))
w26 <- w26 %>% select(h_hidp, h_tenure_dv) %>% mutate(across(everything(), as.character))

w26_ind <- import(paste0(path_data, "ukhls/h_indresp.dta"))
w26_ind <- w26_ind %>% select(pidp, h_hidp, h_age_dv, h_jbbgy, h_gor_dv, h_jlendy) %>% mutate(across(everything(), as.character))

w26 <- w26 %>% left_join(w26_ind, by = "h_hidp")

colnames(w26) <- c("hid", "homeowner", "orgpid", "age", "job_tenure", "region", "last_job")

rm(w26_ind)

# Wave 27 (UKHLS)
w27 <- import(paste0(path_data, "ukhls/i_hhresp.dta"))
w27 <- w27 %>% select(i_hidp, i_tenure_dv) %>% mutate(across(everything(), as.character))

w27_ind <- import(paste0(path_data, "ukhls/i_indresp.dta"))
w27_ind <- w27_ind %>% select(pidp, i_hidp, i_age_dv, i_jbbgy, i_gor_dv, i_jlendy) %>% mutate(across(everything(), as.character))

w27 <- w27 %>% left_join(w27_ind, by = "i_hidp")

colnames(w27) <- c("hid", "homeowner", "orgpid", "age", "job_tenure", "region", "last_job")

rm(w27_ind)

# Wave 28 (UKHLS)
w28 <- import(paste0(path_data, "ukhls/j_hhresp.dta"))
w28 <- w28 %>% select(j_hidp, j_tenure_dv) %>% mutate(across(everything(), as.character))

w28_ind <- import(paste0(path_data, "ukhls/j_indresp.dta"))
w28_ind <- w28_ind %>% select(pidp, j_hidp, j_age_dv, j_jbbgy, j_gor_dv, j_jlendy) %>% mutate(across(everything(), as.character))

w28 <- w28 %>% left_join(w28_ind, by = "j_hidp")

colnames(w28) <- c("hid", "homeowner", "orgpid", "age", "job_tenure", "region", "last_job")

rm(w28_ind)

# Wave 29 (UKHLS)
w29 <- import(paste0(path_data, "ukhls/k_hhresp.dta"))
w29 <- w29 %>% select(k_hidp, k_tenure_dv) %>% mutate(across(everything(), as.character))

w29_ind <- import(paste0(path_data, "ukhls/k_indresp.dta"))
w29_ind <- w29_ind %>% select(pidp, k_hidp, k_age_dv, k_jbbgy, k_gor_dv, k_jlendy) %>% mutate(across(everything(), as.character))

w29 <- w29 %>% left_join(w29_ind, by = "k_hidp")

colnames(w29) <- c("hid", "homeowner", "orgpid", "age", "job_tenure", "region", "last_job")

rm(w29_ind)

# Wave 30 (UKHLS)
w30 <- import(paste0(path_data, "ukhls/l_hhresp.dta"))
w30 <- w30 %>% select(l_hidp, l_tenure_dv) %>% mutate(across(everything(), as.character))

w30_ind <- import(paste0(path_data, "ukhls/l_indresp.dta"), encoding = "UTF-8")
w30_ind <- w30_ind %>% select(pidp, l_hidp, l_age_dv, l_jbbgy, l_gor_dv, l_jlendy) %>% mutate(across(everything(), as.character))

w30 <- w30 %>% left_join(w30_ind, by = "l_hidp")

colnames(w30) <- c("hid", "homeowner", "orgpid", "age", "job_tenure", "region", "last_job")

rm(w30_ind)

# Combine
waves_combined <- bind_rows(w1, w2, w3, w4, w5, w6, w7, w8, w9, w10, w11, w12, w13, w14, w15, w16, w17, w18, w19, w20, w21, w22, w23, w24, w25, w26, w27, w28, w29, w30)

rm(path_data, w1, w2, w3, w4, w5, w6, w7, w8, w9, w10, w11, w12, w13, w14, w15, w16, w17, w18, w19, w20, w21, w22, w23, w24, w25, w26, w27, w28, w29, w30)

waves_combined <- waves_combined %>% select(-c("job_tenure", "last_job"))


```

```{r}
# Adjust housing tenure and job tenure variables
waves_combined[waves_combined < 0] <- NA

waves_combined <- waves_combined %>%
  select(-c("hid")) %>%
  mutate(
    social_housing = case_when(
      homeowner %in% c("3", "4") ~ 1,
      !homeowner %in% c("3", "4") ~ 0),
    homeowner_type = case_when(
      homeowner %in% c("2") ~ "Mortgagor",
      homeowner %in% c("1") ~ "Outright owner",
      !homeowner %in% c("1", "2") ~ "Renter"),
    homeowner = case_when(
      homeowner %in% c("1", "2") ~ "Homeowner",
      !homeowner %in% c("1", "2") ~ "Renter"))

waves_combined <- waves_combined %>%
  mutate(region = case_when(
    region %in% c("1") ~ "North_West",
    region %in% c("2") ~ "North_East",
    region %in% c("3") ~ "Yorkshire_and_the_Humber",
    region %in% c("4") ~ "East_Midlands",
    region %in% c("5") ~ "West_Midlands",
    region %in% c("6") ~ "East_England",
    region %in% c("7") ~ "London",
    region %in% c("8") ~ "South_West",
    region %in% c("9") ~ "East_West",
    region %in% c("10") ~ "Wales",
    region %in% c("11") ~ "Scotland",
    region %in% c("12") ~ "Northern_Ireland",
    T ~ NA_character_)) %>% 
  unique()

waves_combined <- waves_combined %>%
  ungroup() %>%
  group_by(orgpid, age) %>%
  filter(n() == 1) %>%
  ungroup()

data_uk <- data_uk %>% mutate(age = as.character(age)) %>% left_join(waves_combined, by = c("orgpid", "age"))

rm(waves_combined)

```

```{r}
# keep respondents only until they enter retirement
data_uk <- data_uk %>% group_by(orgpid) %>% mutate(retired = cumsum(retired)) %>% filter(retired %in% c(0, 1))

data_uk <- data_uk %>% group_by(orgpid) %>% mutate(retired = cumsum(retired)) %>% filter(retired %in% c(0, 1))

# use self-rated health measure of poor health due to lower missingness than in subjective health
data_uk <- data_uk %>% select(-c(poor_health_sb)) %>% rename("poor_health" = "poor_health_sr")

# drop duplicate rows
data_uk$na <- rowSums(is.na(data_uk))
data_uk <- data_uk %>% group_by(orgpid, age) %>% filter(na == min(na)) %>% slice(1) %>% select(-na)

# drop household identifier and adjust variables types
data_uk <- data_uk %>%
  ungroup() %>%
  mutate(
    orgpid = as.factor(orgpid),
    age = as.numeric(age),
    country = as.factor(country),
    year = as.factor(year),
    female = as.numeric(female),
    year_born = as.numeric(year_born),
    cohort_born = as.factor(cohort_born),
    retired = as.numeric(retired),
    migrant = as.factor(migrant),
    education_level = as.factor(education_level),
    marital_status = as.factor(marital_status),
    partner_in_hh = as.numeric(partner_in_hh),
    children_any = as.numeric(children_any),
    household_size = as.numeric(household_size),
    poor_health = as.numeric(poor_health),
    employment_status = as.factor(employment_status),
    employment_status_partner = as.factor(employment_status_partner),
    hh_income_eq = as.numeric(hh_income_eq),
    occupation_group = as.factor(occupation_group),
    public_sector = as.numeric(public_sector),
    self_employed = as.numeric(self_employed),
    unemployment_rate = as.numeric(unemployment_rate),
    inflation_rate = as.numeric(inflation_rate),
    homeowner = as.factor(homeowner),
    homeowner_type = as.factor(homeowner_type),
    social_housing = as.factor(social_housing),
    region = as.factor(region)) %>% 
  unique()

```

```{r}
save.image("data/data_preprocessed/data_uk.RData")

```

```{r}
```
