==> 2.1_descriptives.Rmd <==
---
title: "4.1_descriptives"
author: "Jan Einhoff"
date: "05/2024"
---

This R markdown file is used to create descriptive statistics.

```{r setup}
# Load packages
library("tidyverse")
library("survminer")
library("survival")
library("pammtools")
library("openxlsx")

```

# Missingness
```{r}
# Clear workspace
rm(list = ls())

# Load results files
load("data/data_final_de.RData")

data <- data_de %>% select(-c("country", "orgpid", "age", "year", "female", "cohort_born", "year_born"))

# Check overall missingness
(sum(is.na(data)) * 100 )/(prod(dim(data)))

# Check missingness in each variable
round(colSums(is.na(data))/nrow(data) * 100, 1)

# Check observations with at least one missing
round(100 - (data %>% na.omit() %>% nrow()) / nrow(data) * 100, 1)

```

# Survival plots
```{r}
rm(list = ls())

load("data/data_final_uk.RData")

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_1 %>% mutate(age1 = age - 1))
surv_1 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_2 %>% mutate(age1 = age - 1))
surv_2 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_3 %>% mutate(age1 = age - 1))
surv_3 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_4 %>% mutate(age1 = age - 1))
surv_4 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_5 %>% mutate(age1 = age - 1))
surv_5 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_6 %>% mutate(age1 = age - 1))
surv_6 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_7 %>% mutate(age1 = age - 1))
surv_7 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_8 %>% mutate(age1 = age - 1))
surv_8 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_9 %>% mutate(age1 = age - 1))
surv_9 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

fit <- survfit(Surv(age1, age, retired) ~ homeowner_type, data = data_uk_imp_10 %>% mutate(age1 = age - 1))
surv_10 <- surv_summary(fit) %>% select(time, homeowner_type, surv, std.err)

survival_combined <- rbind(surv_1, surv_2, surv_3, surv_4, surv_5, surv_6, surv_7, surv_8, surv_9, surv_10)

library("psfmi")

survival_combined <- survival_combined %>%
  group_by(homeowner_type, time) %>% mutate(
    est = mean(surv),
    se = sqrt(mean(std.err^2) + (1 + (1/10)) * var(surv)),
    survival = pool_RR(surv, std.err, conf.level = 0.95, nrow(data_uk_imp_1 %>% filter(age <= age)), 1)[1],
    conf_low = pool_RR(surv, std.err, conf.level = 0.95, nrow(data_uk_imp_1 %>% filter(age <= age)), 1)[2],
    conf_high = pool_RR(surv, std.err, conf.level = 0.95, nrow(data_uk_imp_1 %>% filter(age <= age)), 1)[3]) %>%
  select(homeowner_type, time, est, se, survival, conf_low, conf_high) %>%unique()

survival_combined <- survival_combined %>%
  group_by(homeowner_type) %>%
  complete(time = 50:66) %>%
  arrange(homeowner_type, time) %>%
  fill(1:5, .direction = "downup") %>% unique()

```


```{r}

surv_ho <- survival_combined %>% 
  filter(homeowner_type == "Outright owner") %>%
  mutate(surv_ho = est, se_ho = se) %>%
  ungroup() %>%
  select(time, surv_ho, se_ho)

surv_hm <- survival_combined %>% 
  filter(homeowner_type == "Mortgagor") %>%
  mutate(surv_hm = est, se_hm = se) %>%
  ungroup() %>%
  select(time, surv_hm, se_hm)

surv_r <- survival_combined %>%
  filter(homeowner_type == "Renter") %>%
  mutate(surv_r = est, se_r = se) %>%
  ungroup() %>%
  select(time, surv_r, se_r)

surv <- surv_r %>% left_join(surv_hm, by = "time") %>% left_join(surv_ho, by = "time")

surv <- surv %>%
  mutate(
    diff_hm = (surv_hm - surv_r) * 100,
    se_hm = sqrt(se_hm^2 + se_r^2) * 100,
    lower_hm = diff_hm - 1.96 * se_hm,
    upper_hm = diff_hm + 1.96 * se_hm) %>%
  select(time, diff_hm, se_hm, lower_hm, upper_hm)


surv <- surv %>%
  mutate(
    diff_ho = (surv_ho - surv_r) * 100,
    se_ho = sqrt(se_ho^2 + se_r^2) * 100,
    lower_ho = diff_ho - 1.96 * se_ho,
    upper_ho = diff_ho + 1.96 * se_ho,
    diff_hm = (surv_hm - surv_r) * 100,
    se_hm = sqrt(se_hm^2 + se_r^2) * 100,
    lower_hm = diff_hm - 1.96 * se_hm,
    upper_hm = diff_hm + 1.96 * se_hm) %>%
  select(time, diff_ho, se_ho, lower_ho, upper_ho, diff_hm, se_hm, lower_hm, upper_hm)


surv_plot <- ggplot(survival_combined, aes(x = time - 0.5, y = 1 - survival, group = homeowner_type, fill = homeowner_type)) +
  geom_stepribbon(aes(ymin = 1 - conf_low, ymax = 1 - conf_high), alpha = 0.25) +
  geom_step(aes(color = homeowner_type), size = 0.8) +
  scale_x_continuous("Age", limits = c(50, 66), breaks = seq(50, 65, 1)) +
  scale_y_continuous("Cumulative risk (% retired)", breaks = seq(0, 1, 0.1), label = scales::percent) +
  scale_fill_manual(values = c("darkblue", "darkred", "darkgreen")) +
  scale_color_manual(values = c("darkblue", "darkred", "darkgreen")) +
  theme_classic() +
  coord_cartesian(ylim = c(0,1), xlim = c(51, 65)) +
  theme(axis.text.x = element_text(color = "black"),
        axis.text.y = element_text(color = "black"),
        panel.grid.major.y = element_line(colour = "grey", size = 0.1),
        panel.spacing.x = unit(0, "pt"),
        strip.placement = 'outside',
        strip.background = element_blank(),
        panel.background = element_blank(),
        plot.background = element_blank(),
        strip.text.x = element_text(colour = "black"),
        legend.position=c(.25,.86),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.background = element_rect(fill = 'transparent'),
        legend.title = element_blank())
ggsave("outputs/plots/plot_surviva_type_de.png", height = 5, width = 4, bg = "transparent", surv_plot)

```

# Homeownership rates
```{r}
# Clear workspace
rm(list = ls())

# Load data
data <- read.xlsx("data/homeownership_rates.xlsx")

data_home <- data %>%
  mutate(across(everything(), as.character)) %>%
  pivot_longer(
    cols = `1900`:`2019`, 
    names_to = "year",
    values_to = "homeownership_rate") %>%
  na.omit() %>%
  filter(
    Country %in% c("UK", "Germany"),
    year %in% c(1970:2020)) %>%
  mutate(
    homeownership_rate = as.numeric(homeownership_rate),
    year = as.numeric(year))

home_plot <- ggplot(data_home, aes(x = year, y = homeownership_rate / 100, group = Country, color = Country)) +
  geom_point(size = 1) +
  geom_line() +
  theme_classic() +
  scale_fill_manual(values = c("darkblue", "darkred")) +
  scale_color_manual(values = c("darkblue", "darkred")) +
  scale_x_continuous(breaks = seq(1990, 2020, 5), "Year", labels = c("'90", "'95", "'00", "'05", "'10", "'15", "'20")) +
  scale_y_continuous(breaks = seq(0.2, 1, 0.05), "Home ownership rate", label = scales::percent) +
  coord_cartesian(xlim = c(1990, 2020)) +
  theme(axis.text.x = element_text(color = "black"),
        axis.text.y = element_text(color = "black"),
        panel.grid.major.y = element_line(colour = "grey", size = 0.1),
        panel.spacing.x = unit(0, "pt"),
        strip.placement = 'outside',
        strip.background = element_blank(),
        panel.background = element_blank(),
        plot.background = element_blank(),
        strip.text.x = element_text(colour = "black"),
        legend.position="none",
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.background = element_rect(fill = 'transparent'),
        legend.title = element_blank())
ggsave("outputs/plot_homeownership.png", height = 5, width = 4, bg = "transparent", home_plot)

```

# Retirement age
```{r}
data <- read.xlsx("data/retirement_ages.xlsx")

data_ret <- data %>%
  na.omit() %>%
  filter(Country %in% c("UK", "Germany"),
         year %in% c(1970:2020))

retage_plot <- ggplot(data_ret, aes(x = as.numeric(year), y = ret_age, group = interaction(gender, Country), color = Country, linetype = gender)) +
  geom_point(size = 1) +
  geom_line() +
  theme_classic() +
  scale_fill_manual(values = c("darkblue", "darkred")) +
  scale_color_manual(values = c("darkblue", "darkred")) +
  scale_x_continuous(breaks = seq(1990, 2020, 5), "Year", labels = c("'90", "'95", "'00", "'05", "'10", "'15", "'20")) +
  scale_y_continuous(breaks = seq(20, 100, 1), "Average effective retirement age") +
  coord_cartesian(xlim = c(1990, 2020)) +
  theme(axis.text.x = element_text(color = "black"),
        axis.text.y = element_text(color = "black"),
        panel.grid.major.y = element_line(colour = "grey", size = 0.1),
        panel.spacing.x = unit(0, "pt"),
        strip.placement = 'outside',
        strip.background = element_blank(),
        panel.background = element_blank(),
        plot.background = element_blank(),
        strip.text.x = element_text(colour = "black"),
        legend.position="none",
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.background = element_rect(fill = 'transparent'),
        legend.title = element_blank())
ggsave("outputs/plot_retirementage.png", height = 5, width = 4, bg = "transparent", retage_plot)

```

```{r}
data_home <- data_home %>% mutate(gender = "Men", var = "Homeownership share") %>% select(-c("iso"))

colnames(data_home) <- c("country", "year", "rate", "group", "var")

data_ret <- data_ret %>% mutate(var = "Retirement age")

colnames(data_ret) <- c("group", "year", "country", "rate", "var")

data <- rbind(data_home, data_ret)

data$Gender <- data$group
data$Country <- data$country

legend_plot <- ggplot(data %>% filter(year %in% c(1990:2020)), aes(x = as.numeric(year), y = rate, group = interaction(Gender, Country), color = Country, linetype = Gender)) +
  facet_wrap(~ var, scales = "free_y") +
  geom_point(size = 0.5) +
  geom_line() +
  theme_classic() +
  theme(text = element_text(size = 13,  family = "Arial"), legend.position = "bottom") +
  scale_x_continuous(breaks = seq(1960, 2020, 5), "Year", labels = c("'60", "'65", "'70", "'75", "'80", "'85", "'90", "'95", "'00", "'05", "'10", "'15", "'20")) +
  scale_y_continuous(element_blank()) +
  scale_color_manual(values = c("darkblue", "darkred")) +
  theme(strip.placement = 'outside',
        strip.background = element_blank(),
        panel.background = element_blank(),
        plot.background = element_blank(),
        strip.text.x = element_text(colour = "black"),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.background = element_rect(fill = 'transparent'),
        legend.title = element_blank())
ggsave("outputs/legend.png", width = 7, height = 5, legend_plot)


```

```{r}
```
